+++ src/assistant/index.ts
@@ -3,7 +3,6 @@ export * from "./data.ts";
 export * from "./file.ts";
 export * from "./folder.ts";
 export * from "./module.ts";
-export * from "./packs.ts";
 export * from "./reroll.ts";
 export * from "./socket.ts";
 export * from "./storage.ts";