+++ src/data/spells/01/figment.ts
@@ -1,0 +1,24 @@
+import { Assistant } from "assistant.ts";
+
+export const path = ["Spells", "1st Rank", "Figment"];
+
+export const actions: Assistant.Action[] = [
+    {
+        trigger: "action",
+        predicate: ["item:slug:figment"],
+        process: async (data: Assistant.Data) => {
+            if (!data.speaker) return;
+            if (!data.item?.isOfType("spell")) return;
+
+            await game.assistant.socket.addEffect(
+                data.speaker.actor,
+                PF2E_SPELL_EFFECTS["spell-effect-figment"],
+                {
+                    origin: data.speaker,
+                    item: data.item,
+                    target: data.speaker
+                }
+            );
+        }
+    }
+];